Add MIT/Expat license file
Since early on, this project has declared itself as licensed under the
MIT license.  It's good form to include the actual text of the license
in the repository.  This provides greater clarity about the license
terms and allows various tools to automatically detect the project's
license.

In this commit, we add the text of the MIT/Expat license under whose
terms this project is distributed.