Using back r2d2-sqlite and rusqlite as dependency
Previously this project was using r2d2-sqlite3 and sqlite3 for handling the sqlite platform. But since rusqlite now has exposed the needed api from sqlite, it will now be possible to use it. The sqlite test are now also fixed with a quirk to insert multiple data using the limited single insert of sqlite